fix php errors in games_series, games_upload, individual_author, magazine_review_score, news_approve, news_edit

+++ Website/AtariLegend/admin/games/games_upload.php
@@ -17,13 +17,12 @@
 
 //load all common functions
 include("../includes/common.php"); 
-include("../includes/config.php"); 
 
 
 //****************************************************************************************
 // We wanna add a new download
 //**************************************************************************************** 
-if ($action == 'add_download' )
+if (isset($action) and $action == 'add_download' )
 {
 	require_once('../includes/pclzip.lib.php');
 	
@@ -117,7 +116,7 @@ if ($action == 'add_download' )
 //****************************************************************************************
 // When the update button has been pressed, the file name and comments get updated
 //**************************************************************************************** 
-if ($action == 'update_download')
+if (isset($action) and $action == 'update_download')
 {
 	mysql_query("UPDATE game_download SET cracker='$cracker', supplier='$supplier', screen='$screen', language='$language', trainer='$trainer', legend='$legend', disks='$disks', set_nr='$set_nr', intro='$intro', harddrive='$harddrive', disable='$disable', version='$version', tos='$tos' WHERE game_download_id='$game_download_id'");
 }
@@ -125,7 +124,7 @@ if ($action == 'update_download')
 //****************************************************************************************
 // This is where the file name and comments get deleted
 //**************************************************************************************** 
-if ($action == "delete_download")
+if (isset($action) and $action == "delete_download")
 {
 	mysql_query("DELETE from game_download WHERE game_download_id='$game_download_id'");
 	unlink ("$game_file_path$game_download_id.zip");
@@ -147,54 +146,55 @@ $game_info=mysql_fetch_array ($SQL_GAME_INFO);
 //get some basic game info
 $smarty->assign('game',
 		 array('game_id' => $game_id,
-		 	   'game_name' => $game_info[game_name]));
+		 	   'game_name' => $game_info['game_name']));
 
 //get the existing downloads
 $SQL_DOWNLOADS = mysql_query("SELECT * FROM game_download WHERE game_id='$game_id'")
 			  	 or die ("Error getting download info");		
 
+$nr_downloads = 1;
 while ($downloads=mysql_fetch_array($SQL_DOWNLOADS)) 
 {
 	// first lets create the filenames
 	$filename = "$game_info[game_name]";
 	
-	if ($game_info[game_year]=='') { $filename .= " (19xx)"; } else { $filename .= " ($game_info[game_year])";}
-	if ($game_info[pub_dev_id]!=='') { $filename .= "($game_info[pub_dev_name])"; }
-	if ($downloads[game_ext]=="stx") { $filename .="[pasti]";}
-	if ($downloads[game_ext]=="msa") { $filename .="[MSA]";}
-	if ($downloads[cracker]!=="") { $filename .="[cr $downloads[cracker]]";}
-	if ($downloads[supplier]!=="") { $filename .="[su $downloads[supplier]]";}
-	if ($downloads[screen]!=="") { $filename .="[$downloads[screen]]";}
-	if ($downloads[language]!=="") { $filename .="[$downloads[language]]";}
-	if ($downloads[trainer]!=="") { $filename .="[$downloads[trainer]]";}
-	if ($downloads[legend]=="1") { $filename .="[AL]";}
-	if ($downloads[disks]!=="0" || $downloads[disks]!=="") { $filename .=" Disk $downloads[disks]";}
+	if ($game_info['game_year']=='') { $filename .= " (19xx)"; } else { $filename .= " ($game_info[game_year])";}
+	if ($game_info['pub_dev_id']!=='') { $filename .= "($game_info[pub_dev_name])"; }
+	if ($downloads['game_ext']=="stx") { $filename .="[pasti]";}
+	if ($downloads['game_ext']=="msa") { $filename .="[MSA]";}
+	if ($downloads['cracker']!=="") { $filename .="[cr $downloads[cracker]]";}
+	if ($downloads['supplier']!=="") { $filename .="[su $downloads[supplier]]";}
+	if ($downloads['screen']!=="") { $filename .="[$downloads[screen]]";}
+	if ($downloads['language']!=="") { $filename .="[$downloads[language]]";}
+	if ($downloads['trainer']!=="") { $filename .="[$downloads[trainer]]";}
+	if ($downloads['legend']=="1") { $filename .="[AL]";}
+	if ($downloads['disks']!=="0" || $downloads['disks']!=="") { $filename .=" Disk $downloads[disks]";}
 	$filename .=".zip";
 	
 	$filepath = $game_file_path;
-	$filepath .= $downloads[game_download_id];
+	$filepath .= $downloads['game_download_id'];
 	
 	//convert the date
 	$date = convert_timestamp($downloads['date']);
 	
 	//start filling the smarty object
 	$smarty->append('downloads',
-	    	 array('game_download_id' => $downloads[game_download_id],
-			 	   'cracker' => $downloads[cracker],
-				   'supplier' => $downloads[supplier],
-				   'intro' => $downloads[intro],
-				   'harddrive' => $downloads[harddrive],
-				   'pal_ntsc' => $downloads[screen],
-				   'language' => $downloads[language],
-				   'trainer' => $downloads[trainer],
-				   'disks' => $downloads[disks],
-				   'set_nr' => $downloads[set_nr],
-				   'legend' => $downloads[legend],
-				   'disable' => $downloads[disable],
+	    	 array('game_download_id' => $downloads['game_download_id'],
+			 	   'cracker' => $downloads['cracker'],
+				   'supplier' => $downloads['supplier'],
+				   'intro' => $downloads['intro'],
+				   'harddrive' => $downloads['harddrive'],
+				   'pal_ntsc' => $downloads['screen'],
+				   'language' => $downloads['language'],
+				   'trainer' => $downloads['trainer'],
+				   'disks' => $downloads['disks'],
+				   'set_nr' => $downloads['set_nr'],
+				   'legend' => $downloads['legend'],
+				   'disable' => $downloads['disable'],
 				   'filename' => $filename,
 				   'filepath' => $filepath,
-				   'version' => $downloads[version],
-				   'tos' => $downloads[tos],
+				   'version' => $downloads['version'],
+				   'tos' => $downloads['tos'],
 				   'date' => $date));
 
 	$nr_downloads++;
@@ -202,7 +202,7 @@ while ($downloads=mysql_fetch_array($SQL_DOWNLOADS))
 
 $smarty->assign('nr_downloads',$nr_downloads);
 
-$smarty->assign("user_id",$_SESSION[user_id]);
+$smarty->assign("user_id",$_SESSION['user_id']);
 $smarty->assign('games_upload_tpl', '1');
 
 //Send all smarty variables to the templates
